+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Card from "./components/card/Card";
-import AddFormModal from "./components/card/AddFormModal";
-import Header from "./components/card/Header";
+import AddFormModal from "./components/Modal/AddFormModal";
+import Header from "./components/Header/Header";
 import "./App.css";
 
 const App = () => {
